[#47] 전체 코드 리팩토링

+++ bus-seat/src/main/java/org/kernel360/busseat/route_station/dto/RouteStationDto.java
@@ -18,7 +18,7 @@ public class RouteStationDto {
 
 	private Long stationId;
 
-	private Short stationSeq;
+	private Integer stationSeq;
 
 	private String stationName;
